Enable SSO debug logs (#23)
I've read Janrain's sso source and figured out that there is a way to
enable debugging output. Weirdly enough it seems to fixing #22 for
me (both in Firefox and Chrome).

I'm not really trusting such fix, but I thought that it's a good
idea to commit it and try on staging. It won't hurt to have debug
logs either way.