Add elections to deploy.rb, restructure env_variables (#64)
I only added `elections` to a list of `apps`. The script would have
failed if any of the `env_variables` weren't defined. We don't need
any of existing variables in the `elections`, but we'll soon need some
other ones (the backend URL I suppose).

Instead of having a list of `apps` and a global list of
`env_variables` I made an object where app names are keys
and their env variables are in the values (possibly with
some other configuration in the future).

TIL ruby, WAAAAAT :D