Add pre-commit hook
This is for the monorepo structure and Netlify builds.

As the monorepo workflow in Netlify work so that it checks versioned
changes under an app directory, it's not complete for our purposes.
In Purescript applications, we also need to see if the build
directory (usually `output`) has some changes in it, and
trigger the deploy if changes are found. This is done with the
`.trigger-deploy` file, which simply includes the size of the build
directory. An amazing perl script has been added just for this
purpose, and to make it somewhat automatic, it's useful
to have it as a git pre-commit hook.

More info https://www.netlify.com/blog/2019/10/09/launching-monorepo-support-for-netlify-sites/