Add hashes of purescript dependencies (#252)
This is for the monorepo structure and Netlify builds.

As the monorepo workflow in Netlify work so that it checks versioned
changes under an app directory, it's not complete for our purposes.
In Purescript applications, we also need to see if the build
directory (usually `output`) has some changes in it, and
trigger the deploy if changes are found. This is done with the
`.build-size-sha256` file, which simply includes hashed content of the build
directory. An amazing perl script has been added just for this
purpose, and to make it somewhat automatic, it's useful
to have it as a git pre-commit hook.